backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-multipart
pyjwt
passlib[bcrypt]